Added arrays for configurations.
Added ability to save and load configurations with arrays (both of primitives and of objects). Any primitive is easily loaded, whereas object types require another ConfigBase to be defined.

I personally think this addition is pretty badass. This means that we don't have to be super sad about the lack of reflection and serialization in C++ because we have the power of macros (as limited as that may be).