Added basic python parsing.
This will change heavily in the future, but for now it is doing the job. A lot of work needs to be done to figure this out.

Right now the parser will look for java source, extract it, and begin to parse it. There is still a lot of stuff that needs to be done to make sure we have a good representation of java.